requirements.txt
fastapi
uvicorn
python-dotenv
amadeus
langchain-ollama
langchain-core
langgraph
streamlit
pydantic
requests
beautifulsoup4
streamlit